roqoqo-aqt/tests/integration/backend.rs
// Copyright © 2021-2023 HQS Quantum Simulations GmbH. All Rights Reserved.
//
// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
// in compliance with the License. You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software distributed under the
// License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
// express or implied. See the License for the specific language governing permissions and
// limitations under the License.

use roqoqo::prelude::*;
use roqoqo::registers::BitRegister;
use roqoqo::{operations::*, Circuit};
use roqoqo_aqt::Backend;
use roqoqo_aqt::{devices::AqtDevice, AqtApi};
use roqoqo_test::prepare_monte_carlo_gate_test;
use serde_json::{json, Value};
use std::collections::HashMap;
use std::env;
use tokio::task::spawn_blocking;
use wiremock::matchers::{body_json, method, path};
use wiremock::{Mock, MockServer, ResponseTemplate};

#[test]
fn init_backend() {
    if env::var("AQT_ACCESS_TOKEN").is_ok() {
        let device = AqtDevice::new(2);
        let _backend = Backend::new(device, None).unwrap();
    } else {
        let device = AqtDevice::new(2);
        let ok = Backend::new(device, None).is_err();
        assert!(ok);
        let device = AqtDevice::new(2);
        let ok = Backend::new(device, Some("dummy_access_token".to_string())).is_ok();
        assert!(ok);
    }
}

// Test to_aqt_json function of Backend
#[test]
fn test_to_aqt_json() {
    let json_aqt_instructions = json!([
    {
        "operation": "R",
        "phi": 0.0,
        "qubit": 0,
        "theta": 1.0,
    },
    {
        "operation": "RZ",
        "phi": 1.0,
        "qubit": 0,
    },
    {
        "operation": "RZ",
        "phi": 1.0,
        "qubit": 1,
    },
    {
        "operation": "R",
        "phi": 0.5,
        "qubit": 1,
        "theta": 1.0,
    },
    {
        "operation": "RXX",
        "qubits": [
          0,
          1
        ],
        "theta": 0.5
    },
    {
        "operation": "RXX",
        "qubits": [
          0,
          1
        ],
        "theta": 0.5
    },
    {
        "operation": "MEASURE"
    }
    ]);

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 2, true);
    circuit += PauliX::new(0);
    circuit += PauliZ::new(0);
    circuit += PauliZ::new(1);
    circuit += PauliY::new(1);
    circuit += MolmerSorensenXX::new(0, 1);
    circuit += VariableMSXX::new(0, 1, 1.0.into());
    circuit += MeasureQubit::new(1, "ro".to_string(), 1);

    let backend = Backend::new(AqtDevice::new(1), Some("dummy".into())).unwrap();
    let aqt_json_string = backend.to_aqt_json(circuit.iter()).unwrap();
    let json_val: Value = serde_json::from_str(&aqt_json_string).unwrap();
    assert_eq!(json_val, json_aqt_instructions)
}

// Test API endpoint calls and error handling for erroneous status codes.
#[tokio::test]
async fn api_status_test() {
    let aqt_run_response_empty = json!({"job": {},
        "response": {
        "status": ""
      }
    });
    #[derive(Clone)]
    struct MockAqtDevice {
        pub number_qubits: usize,
        pub mock_host: String,
    }

    impl AqtApi for MockAqtDevice {
        fn remote_host(&self) -> String {
            self.mock_host.to_string()
        }

        fn number_qubits(&self) -> usize {
            self.number_qubits
        }

        fn is_https(&self) -> bool {
            false
        }

        fn id(&self) -> String {
            "dummy".to_string()
        }
    }

    let server = MockServer::start().await;
    let uri = server.uri();
    // matching expected post body
    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_empty))
        .mount(&server)
        .await;

    let mut circuit = Circuit::new();
    circuit += PauliX::new(0);
    let client = spawn_blocking(move || reqwest::blocking::Client::builder().build())
        .await
        .unwrap()
        .unwrap();

    let mock_device = MockAqtDevice {
        number_qubits: 1,
        mock_host: format!("{}/mock/", uri),
    };
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let (data, _, _) = backend
        .convert_circuit_to_aqt_instructions(circuit.iter())
        .unwrap();
    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.post_job(&client_cloned, data))
        .await
        .unwrap();
    assert!(res.is_ok());

    server.verify().await;
    server.reset().await;

    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .respond_with(ResponseTemplate::new(404))
        .mount(&server)
        .await;

    let (data, _, _) = backend
        .convert_circuit_to_aqt_instructions(circuit.iter())
        .unwrap();
    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.post_job(&client_cloned, data))
        .await
        .unwrap();
    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::NetworkError {
        msg: "Failed to post job to server. Request to server failed with HTTP status code 404"
            .to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;

    Mock::given(method("GET"))
        .and(path("/mock/result/dummy_id"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_empty))
        .mount(&server)
        .await;

    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.get_result(&client_cloned, "dummy_id"))
        .await
        .unwrap();

    assert!(res.is_ok());

    server.verify().await;
    server.reset().await;

    Mock::given(method("GET"))
        .and(path("/mock/result/dummy_id"))
        .respond_with(ResponseTemplate::new(404))
        .mount(&server)
        .await;

    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.get_result(&client_cloned, "dummy_id"))
        .await
        .unwrap();

    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::NetworkError {
        msg: "Failed to get result from server. Request to server failed with HTTP status code 404"
            .to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;
}

// Test request body for api requests
#[tokio::test]
async fn api_request_body_test() {
    let aqt_expect_post_body_small = json!({
        "job_type": "quantum_circuit",
        "label": "qoqo_aqt_backend",
        "payload": {
            "circuits": [
            {
                "number_of_qubits": 1,
                "quantum_circuit": [
                {
                    "operation": "R",
                    "phi": 0.0,
                    "qubit": 0,
                    "theta": 1.0,
                },
                {
                    "operation": "MEASURE"
                }
                ],
                "repetitions": 5
            }
            ]
        }
        }
    );

    let aqt_expect_post_body = json!({
        "job_type": "quantum_circuit",
        "label": "qoqo_aqt_backend",
        "payload": {
            "circuits": [
            {
                "number_of_qubits": 2,
                "quantum_circuit": [
                {
                    "operation": "R",
                    "phi": 0.0,
                    "qubit": 0,
                    "theta": 1.0,
                },
                {
                    "operation": "RZ",
                    "phi": 1.0,
                    "qubit": 0,
                },
                {
                    "operation": "RZ",
                    "phi": 1.0,
                    "qubit": 1,
                },
                {
                    "operation": "R",
                    "phi": 0.5,
                    "qubit": 1,
                    "theta": 1.0,
                },
                {
                    "operation": "RXX",
                    "qubits": [
                      0,
                      1
                    ],
                    "theta": 0.5
                },
                {
                    "operation": "RXX",
                    "qubits": [
                      0,
                      1
                    ],
                    "theta": 0.5
                },

                {
                    "operation": "MEASURE"
                }
                ],
                "repetitions": 5
            }
            ]
        }
        }
    );

    let aqt_run_response_empty = json!({"job": {
          "job_id": "dummy_test_id",
      },
      "response": {
      "status": "queued"
    }
      });

    #[derive(Clone)]
    struct MockAqtDevice {
        pub number_qubits: usize,
        pub mock_host: String,
    }

    impl AqtApi for MockAqtDevice {
        fn remote_host(&self) -> String {
            self.mock_host.to_string()
        }

        fn number_qubits(&self) -> usize {
            self.number_qubits
        }

        fn is_https(&self) -> bool {
            false
        }

        fn id(&self) -> String {
            "dummy".to_string()
        }
    }

    let server = MockServer::start().await;
    let uri = server.uri();
    // matching expected post body
    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .and(body_json(&aqt_expect_post_body_small))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_empty))
        .mount(&server)
        .await;

    let client = spawn_blocking(move || reqwest::blocking::Client::builder().build())
        .await
        .unwrap()
        .unwrap();

    let mock_device = MockAqtDevice {
        number_qubits: 1,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let (data, _, _) = backend
        .convert_circuit_to_aqt_instructions(circuit.iter())
        .unwrap();
    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.post_job(&client_cloned, data))
        .await
        .unwrap();
    assert!(res.is_ok());
    assert_eq!(res.unwrap().status(), "queued");

    server.verify().await;
    server.reset().await;

    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .and(body_json(&aqt_expect_post_body))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_empty))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 2,
        mock_host: format!("{}/mock/", uri),
    };
    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 2, true);
    circuit += PauliX::new(0);
    circuit += PauliZ::new(0);
    circuit += PauliZ::new(1);
    circuit += PauliY::new(1);
    circuit += MolmerSorensenXX::new(0, 1);
    circuit += VariableMSXX::new(0, 1, 1.0.into());
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let (data, _, _) = backend
        .convert_circuit_to_aqt_instructions(circuit.iter())
        .unwrap();
    let backend_cloned = backend.clone();
    let client_cloned = client.clone();
    let res = spawn_blocking(move || backend_cloned.post_job(&client_cloned, data))
        .await
        .unwrap();
    assert!(res.is_ok());
    let res = res.unwrap();
    assert_eq!(res.status(), "queued");
    assert_eq!(res.job_id(), "dummy_test_id");

    server.verify().await;
    server.reset().await;
}

#[tokio::test]
async fn api_resources_error_mock_test() {
    let aqt_resouce_details_offline = json!({
        "id": "dummy",
        "name": "Noisy Simulator",
        "type": "simulator",
        "status": "offline",
        "available_qubits": 12
    });
    let aqt_resouce_details_low_qubits = json!({
        "id": "dummy",
        "name": "Noisy Simulator",
        "type": "simulator",
        "status": "online",
        "available_qubits": 2
    });

    #[derive(Clone)]
    struct MockAqtDevice {
        pub number_qubits: usize,
        pub mock_host: String,
    }

    impl AqtApi for MockAqtDevice {
        fn remote_host(&self) -> String {
            self.mock_host.to_string()
        }

        fn number_qubits(&self) -> usize {
            self.number_qubits
        }

        fn is_https(&self) -> bool {
            false
        }

        fn id(&self) -> String {
            "dummy".to_string()
        }
    }

    let server = MockServer::start().await;
    let uri = server.uri();

    Mock::given(method("GET"))
        .and(path("/mock/resources/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_resouce_details_offline))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 2,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let res = spawn_blocking(move || backend.run_circuit(&circuit))
        .await
        .unwrap();
    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::NetworkError {
        msg: "AQT resource is currently ofline".to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;

    Mock::given(method("GET"))
        .and(path("/mock/resources/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_resouce_details_low_qubits))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 10,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let res = spawn_blocking(move || backend.run_circuit(&circuit))
        .await
        .unwrap();
    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::GenericError {
        msg: "Insuffient qubits on backend device. Maximum available qubits: 2.".to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;
}
// Test backend run function with a mock device
#[tokio::test]
async fn api_backend_mock_test() {
    let aqt_resouce_details_online = json!({
      "id": "dummy",
      "name": "Noisy Simulator",
      "type": "simulator",
      "status": "online",
      "available_qubits": 12
    });
    let aqt_expect_post_body_small = json!({
        "job_type": "quantum_circuit",
        "label": "qoqo_aqt_backend",
        "payload": {
            "circuits": [
            {
                "number_of_qubits": 1,
                "quantum_circuit": [
                {
                    "operation": "R",
                    "phi": 0.0,
                    "qubit": 0,
                    "theta": 1.0,
                },
                {
                    "operation": "MEASURE"
                }
                ],
                "repetitions": 5
            }
            ]
        }
        }
    );

    let aqt_run_response_queued = json!({"job": {
          "job_id": "dummy_test_id",
      },
      "response": {
      "status": "queued"
    }
      });

    let aqt_run_response_finished = json!({
      "job": {
        "job_type": "quantum_circuit",
        "label": "",
        "job_id": "dummy_test_id",
      },
      "response": {
        "status": "finished",
        "result": {
          "0": [
            [
              1
            ],
            [
              1
            ],
            [
              1
            ],
            [
              1
            ],
            [
              1
            ],
          ]
        }
      }
    });

    let aqt_run_response_failed = json!({
     "job": {
        "job_type": "quantum_circuit",
        "label": "",
        "job_id": "dummy_test_id",
      },
      "response": {
        "message": "detailed error message",
        "status": "error"
      }
    });

    let aqt_run_response_cancelled = json!({
      "job": {
        "job_type": "quantum_circuit",
        "label": "",
        "job_id": "dummy_test_id",
        "resource_id": "",
        "workspace_id": ""
      },
      "response": {
        "status": "cancelled"
      }
    });

    #[derive(Clone)]
    struct MockAqtDevice {
        pub number_qubits: usize,
        pub mock_host: String,
    }

    impl AqtApi for MockAqtDevice {
        fn remote_host(&self) -> String {
            self.mock_host.to_string()
        }

        fn number_qubits(&self) -> usize {
            self.number_qubits
        }

        fn is_https(&self) -> bool {
            false
        }

        fn id(&self) -> String {
            "dummy".to_string()
        }
    }

    let server = MockServer::start().await;
    let uri = server.uri();

    Mock::given(method("GET"))
        .and(path("/mock/resources/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_resouce_details_online))
        .mount(&server)
        .await;

    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .and(body_json(&aqt_expect_post_body_small))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_queued))
        .mount(&server)
        .await;

    Mock::given(method("GET"))
        .and(path("/mock/result/dummy_test_id"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_finished))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 1,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let res = spawn_blocking(move || backend.run_circuit(&circuit))
        .await
        .unwrap();
    assert!(res.is_ok());
    let (bit_registers, _, _) = res.unwrap();
    let mut expected_br = HashMap::<String, Vec<BitRegister>>::new();
    expected_br.insert(
        "ro".to_string(),
        vec![vec![true], vec![true], vec![true], vec![true], vec![true]],
    );
    assert_eq!(bit_registers, expected_br);

    server.verify().await;
    server.reset().await;

    Mock::given(method("GET"))
        .and(path("/mock/resources/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_resouce_details_online))
        .mount(&server)
        .await;

    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .and(body_json(&aqt_expect_post_body_small))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_queued))
        .mount(&server)
        .await;

    Mock::given(method("GET"))
        .and(path("/mock/result/dummy_test_id"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_cancelled))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 1,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let res = spawn_blocking(move || backend.run_circuit(&circuit))
        .await
        .unwrap();
    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::NetworkError {
        msg: "AQT network backend reported that the job was cancelled".to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;

    Mock::given(method("GET"))
        .and(path("/mock/resources/dummy"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_resouce_details_online))
        .mount(&server)
        .await;

    Mock::given(method("POST"))
        .and(path("/mock/submit/qoqo-integration/dummy"))
        .and(body_json(&aqt_expect_post_body_small))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_queued))
        .mount(&server)
        .await;

    Mock::given(method("GET"))
        .and(path("/mock/result/dummy_test_id"))
        .respond_with(ResponseTemplate::new(200).set_body_json(&aqt_run_response_failed))
        .mount(&server)
        .await;

    let mock_device = MockAqtDevice {
        number_qubits: 1,
        mock_host: format!("{}/mock/", uri),
    };

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let backend = Backend::new(mock_device, Some("DummyAccessToken".to_string())).unwrap();
    let res = spawn_blocking(move || backend.run_circuit(&circuit))
        .await
        .unwrap();
    assert!(res.is_err());
    let e = res.err().unwrap();
    let expected_error = RoqoqoBackendError::NetworkError {
        msg: "AQT network backend reported error: detailed error message".to_string(),
    };
    assert_eq!(e, expected_error);

    server.verify().await;
    server.reset().await;
}

// Test backend run on AQT simulator with small circuit
#[test]
fn api_backend_test_small() {
    let device = AqtDevice { number_qubits: 1 };
    let backend = Backend::new(device, None).unwrap();

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += PragmaRepeatedMeasurement::new("ro".to_string(), 5, None);
    let (bit_registers, _float_registers, _complex_registers) =
        backend.run_circuit(&circuit).unwrap();

    let mut expected_br = HashMap::<String, Vec<BitRegister>>::new();
    expected_br.insert(
        "ro".to_string(),
        vec![vec![true], vec![true], vec![true], vec![true], vec![true]],
    );
    assert_eq!(bit_registers, expected_br);
}

// Test backend run on AQT simulator with small circuit
#[test]
fn api_backend_test_small_two() {
    let device = AqtDevice { number_qubits: 1 };
    let backend = Backend::new(device, None).unwrap();

    let mut circuit = Circuit::new();
    circuit += DefinitionBit::new("ro".to_string(), 1, true);
    circuit += PauliX::new(0);
    circuit += MeasureQubit::new(0, "ro".to_string(), 0);
    let (bit_registers, _float_registers, _complex_registers) =
        backend.run_circuit(&circuit).unwrap();

    let mut expected_br = HashMap::<String, Vec<BitRegister>>::new();
    expected_br.insert("ro".to_string(), vec![vec![true]]);
    assert_eq!(bit_registers, expected_br);
}

/// Test full gate with stochastic application of gates, ignore normally because of length and load
#[test]
#[ignore = "Takes too long and puts large load on AQT servers"]
fn test_full_simple_gate() {
    if env::var("AQT_ACCESS_TOKEN").is_ok() {
        let gate: GateOperation = PauliX::new(0).into();
        let preparation_gates: Vec<SingleQubitGateOperation> = vec![
            PauliX::new(0).into(),
            PauliY::new(0).into(),
            PauliZ::new(0).into(),
        ];
        let basis_rotation_gates: Vec<SingleQubitGateOperation> = vec![
            PauliX::new(0).into(),
            RotateX::new(0, std::f64::consts::FRAC_PI_2.into()).into(),
            RotateZ::new(0, std::f64::consts::FRAC_PI_2.into()).into(),
        ];
        let (measurement, exp_vals) = prepare_monte_carlo_gate_test(
            gate,
            preparation_gates,
            basis_rotation_gates,
            None,
            5,
            200,
        );

        let device = AqtDevice::new(2);
        let backend = Backend::new(device, None).unwrap();
        let measured_exp_vals = backend.run_measurement(&measurement).unwrap().unwrap();
        for (key, val) in exp_vals.iter() {
            assert!((val - measured_exp_vals.get(key).unwrap()).abs() < 1.0);
        }
    }
}
